upload signup component, update signin component, and add routers
