+++ product management system/src/App.jsx
@@ -1,14 +1,19 @@
 import './App.css'
-import SignIn from './Pages/SignIn.jsx'
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
+import SignIn from './Pages/SignIn/SignIn.jsx'
+import SignUp from './Pages/SignUp/SignUp.jsx'
 
 function App() {
 
   return (
     <>
-      <div>
+      <Router>
+        <Routes>
+          <Route path='/signin' element={<SignIn />} />
+          <Route path='/signup' element={<SignUp />} />
+        </Routes>
+      </Router>
 
-        <SignIn />
-      </div>
     </>
   )
 }
