--- product management system/src/Pages/SignIn.css
@@ -1,75 +1,0 @@
-.sign-in-container {
-    display: flex;
-    flex-direction: column;
-    justify-items: center;
-    height: 50vh;
-    width: 50vw;
-    min-height: 400px;
-    min-width: 300px; 
-    max-width: 600px; 
-    border-radius: 10px;
-    background-color: white;
-    box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
-    padding: 20px;
-    margin: 0 auto;
-}
-
-.sign-in-container h1{
-    text-align: center;
-}
-
-.sign-in-container label {
-    display: block;
-    margin-bottom: 5px;
-}
-
-.sign-in-container input {
-  width: 90%;
-  border: 1px solid #d6d5d5;
-  border-radius: 3px;
-  padding: 15px;
-  margin-bottom: 15px;
-}
-
-
-.password-container {
-    display:flex;
-    flex-direction:column;
-}
-
-.sign-in-container .sign-in-button {
-    background-color: #5b6af5;
-    color: white;
-    border: none;
-    width: 99%;
-    border-radius: 3px;
-    padding: 10px;
-    margin-top: 15px;
-}
-
-
-.password-input-container {
-    position: relative;
-}
-
-.password-show-button {
-    position: absolute;
-    right: 10px;
-    top: 40%;
-    transform: translateY(-50%);
-    border: none;
-    
-    background-color: white;
-}
-
-.password-show-button:hover {
-    text-decoration: underline;
-}
-
-
-
-a {
-    font-weight: 500;
-    color: #646cff;
-    text-decoration: underline;
-}
