+++ product management system/src/Pages/SignIn/SignIn.css
@@ -1,0 +1,102 @@
+.sign-in-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+}
+
+.sign-in-form-container {
+  display: flex;
+  flex-direction: column;
+  justify-items: center;
+  width: 50vw;
+  min-height: 430px;
+  min-width: 300px; 
+  max-width: 600px; 
+  border-radius: 10px;
+  background-color: white;
+  box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
+  padding: 20px;
+  margin: 0 auto;
+}
+
+.sign-in-form-container h1{
+  text-align: center;
+}
+
+.sign-in-form-container label {
+  display: block;
+  margin-bottom: 5px;
+}
+
+.sign-in-form-container input {
+  width: calc(100% - 35px);
+  border: 1px solid #d6d5d5;
+  border-radius: 3px;
+  padding: 15px;
+  margin-bottom: 15px;
+}
+
+.password-container {
+  display:flex;
+  flex-direction:column;
+}
+
+.sign-in-form-container .sign-in-button {
+  background-color: #5b6af5;
+  color: white;
+  border: none;
+  width: 100%;
+  border-radius: 3px;
+  padding: 10px;
+  margin-top: 15px;
+}
+
+.password-input-container {
+  position: relative;
+}
+
+.password-show-button {
+  position: absolute;
+  right: 4%;
+  top: 40%;
+  transform: translateY(-50%);
+  border: none;
+  background-color: white;
+}
+
+.password-show-button:hover {
+  text-decoration: underline;
+}
+
+a {
+  font-weight: 500;
+  color: #646cff;
+  text-decoration: underline;
+}
+
+.additional-links {
+  display: flex;
+  flex-direction: row;
+  margin: 10px 0 10px;
+}
+
+.additional-links .reset-password{
+  margin: 1em 0 1em auto;
+}
+
+
+@media (max-width: 800px) {
+  .sign-in-form-container h1{
+    font-size: x-large;
+  }
+
+  .additional-links {
+    flex-direction: column;
+    margin: 10px auto;
+  }
+
+  .additional-links .reset-password{
+    margin: auto;
+  }
+}
