+++ product management system/src/Pages/SignUp/SignUp.css
@@ -1,0 +1,93 @@
+.sign-up-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+}
+
+.sign-up-form-container {
+  display: flex;
+  flex-direction: column;
+  justify-items: center;
+  width: 50vw;
+  min-height: 430px;
+  min-width: 300px; 
+  max-width: 600px; 
+  border-radius: 10px;
+  background-color: white;
+  box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
+  padding: 20px;
+  margin: 0 auto;
+}
+
+.sign-up-form-container h1{
+  text-align: center;
+}
+
+.sign-up-form-container label {
+  display: block;
+  margin-bottom: 5px;
+}
+
+.sign-up-form-container input {
+  width: calc(100% - 35px);
+  border: 1px solid #d6d5d5;
+  border-radius: 3px;
+  padding: 15px;
+  margin-bottom: 15px;
+}
+
+.invalid-container {
+  display:flex;
+  flex-direction:column;
+}
+
+.invalid-container input{
+  border: 1px solid red;
+  margin-bottom: 0;
+}
+
+.invalid-container input:focus {
+  border: 2px solid red;
+  outline: none; 
+}
+
+.invalid-warning{
+  color: red;
+  margin: 0 0 0 auto;
+  font-size: small;
+}
+
+.sign-up-form-container .create-account-button {
+  background-color: #5b6af5;
+  color: white;
+  border: none;
+  width: 100%;
+  border-radius: 3px;
+  padding: 10px;
+  margin-top: 15px;
+}
+
+a {
+  font-weight: 500;
+  color: #646cff;
+  text-decoration: underline;
+}
+
+.additional-links {
+  display: flex;
+  flex-direction: row;
+  margin: 10px 0 10px;
+}
+
+
+@media (max-width: 800px) {
+  .sign-up-form-container h1{
+    font-size: x-large;
+  }
+
+  .additional-links {
+    flex-direction: column;
+    margin: 10px auto;
+  }
+}
